+++ Changelog.md
@@ -5,6 +5,10 @@
 
 1. Deterministic wind plugin:
     * [Pull request 157](https://bitbucket.org/osrf/vrx/pull-requests/157)
+    
+1. Waypoint marker visualization:
+    * [Pull request 162](https://bitbucket.org/osrf/vrx/pull-requests/162)
+    
 
 ### VRX 1.1